CCMAIL: kde-imaging@kde.org As a conclusion to the path issues I've now implemented the following changes in libkipi:
In ImageCollection added path() and uploadPath()
These two function must of course be implemented in the inherited version
of ImageCollectionShared, but a default implementation is given if you do
not care.

path() should return a path for the images. If that does not make sense the
host application may return a common root or even an empty URL. The plugins
are in the documentation of the method suggested to look at the feature
KIPI::AlbumEQDir before expecting that the path is the path of the album.

uploadPath() specifies a directory to place new images. By placing this
method in this class, it makes it possible for host application with
Dir==Album to place new images in these albums, and other application may
choose to place new images on common roots or even in one static directory.

A new KIPI::Features has been added, namely AcceptNewImages, which tells
the plugin if the image accept new images at all. Plugins should never call
uploadPath() of this feature is not present. Likewise if this feature is
present, the host application must implement the uploadPath() method.

In KIPI::Interface a new method has been created, namely addImage( KURL ),
with which the plugin must tell the host application about new images.

svn path=/trunk/kdeextragear-libs-1/libkipi/; revision=312918