CCMAIL: kde-imaging@kde.org
Next try at solving the issue with uploading images.
I've now created a new widget (UploadWidget), which can be used to ask for
a directory to upload to (upload say a scan image from acquireimage plugin,
or a new image converted in batchprocessing plugin).

The upload widget is a directory browser knowing about KURL's.
To let the browser start somewhere, I've added a new method
ImageCollection::uploadRoot.

This isn't 100% complete yet, but I want to show it to Aurelien ;-)

The acquireimage plugin has been changed to used this new class.

I'd esp. like to get some input from the digikam team, as this "breaks"
their work with albums, but on the other hand when they change to recursive
albums, the current scheme will not work anyway.

Cheers
Jesper.

svn path=/trunk/kdeextragear-libs-1/libkipi/; revision=313931