CCMAIL: kde-imaging@kde.org, dfaure@kde.org
Together with David Faure I did yesterday make the necessary changes to
allow icons local to each plugin.

The basic point of all this was that the KIPI::Plugin::actionCollection()
needed to know about the plugins KInstance - as it was till this change, it
got the KInstance for the host application, and would thus look for icons
in the host applications directory.

Making this change did, however, reveal a conflict, namely that each and
every action in the plugin must have the actionCollection as parent -
which was not the case for items in KMenuActions.

The way the host application finds actions right now is by querying the
actionCollection(), which with the above change would also reveal those
actions from menus.

I therefore needed to add another way for the host application to get
information about actions. I added a method called KIPI::Plugin::actions()
which the host application must call to get information about actions. At
the same time I made the actionCollection() method protected, so the host
application can't get to that anymore.

To tell KIPI about actions, then plugins must now call
KIPI::Plugin::addAction() on each action they want to be in the topmost
menu (thus not actions located in KActionMenu's).

Let me summerize the changes
- Host application now need to call actions() rather than
  actionCollection() to get information about action in the plugin.
- plugins now need to give a pointer to KGenericFactory<..>::instance() in
  the constructor of the Plugin superclass.
- the plugin must call addAction() for each toplevel action.

And now to answer the question on how to get local icons.
In Makefile.am, you need a line similar to this:
kipibatchprocessimagesicondir = $(kde_datadir)/kipiplugin_batchprocessimages/icons

The name specified above (the "kipiplugin_batchprocessimages"), must be
the same as what you specify to KGenericFactory in the plugin:

typedef KGenericFactory<Plugin_BatchProcessImages> Factory;
K_EXPORT_COMPONENT_FACTORY( kipiplugin_batchprocessimages,
                            Factory("kipiplugin_batchprocessimages"));

Cheers
Jesper

svn path=/trunk/kdeextragear-libs-1/libkipi/; revision=315552