CC-MAIL:kde-imaging@kde.org
As an action may be located in several windows (e.g. in several image
viewers), the actions are now setup in a setup() method which as argument
take a widget which is the widget the accelerators of the actionCollection
should work on.

actionCollection() is now again public, so it is possible to configure
keybindings.

Host application must now call Plugin::setup( QWidget* ) after having
loaded the plugins.

svn path=/trunk/kdeextragear-libs-1/libkipi/; revision=316558